requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   google-genai
    #   httpx
asgiref==3.11.0
    # via
    #   django
    #   django-cors-headers
    #   strawberry-graphql-django
beautifulsoup4==4.14.3
    # via yfinance
certifi==2026.1.4
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   curl-cffi
charset-normalizer==3.4.4
    # via requests
cross-web==0.4.1
    # via strawberry-graphql
cryptography==46.0.7
    # via google-auth
curl-cffi==0.13.0
    # via yfinance
distro==1.9.0
    # via google-genai
django==5.2.10
    # via
    #   computational-finance (pyproject.toml)
    #   django-cors-headers
    #   strawberry-graphql-django
django-cors-headers==4.9.0
    # via computational-finance (pyproject.toml)
et-xmlfile==2.0.0
    # via openpyxl
frozendict==2.4.7
    # via yfinance
google-auth==2.49.2
    # via google-genai
google-genai==1.72.0
    # via computational-finance (pyproject.toml)
graphql-core==3.2.7
    # via strawberry-graphql
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
multitasking==0.0.12
    # via yfinance
numpy==2.4.1
    # via
    #   computational-finance (pyproject.toml)
    #   pandas
    #   yfinance
openpyxl==3.1.5
    # via computational-finance (pyproject.toml)
packaging==26.0
    # via strawberry-graphql
pandas==3.0.0
    # via
    #   computational-finance (pyproject.toml)
    #   yfinance
peewee==3.19.0
    # via yfinance
platformdirs==4.5.1
    # via yfinance
protobuf==6.33.4
    # via yfinance
psycopg2-binary==2.9.11
    # via computational-finance (pyproject.toml)
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.13.0
    # via google-genai
pydantic-core==2.46.0
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   strawberry-graphql
pytz==2025.2
    # via yfinance
requests==2.32.5
    # via
    #   computational-finance (pyproject.toml)
    #   google-auth
    #   google-genai
    #   yfinance
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via google-genai
soupsieve==2.8.3
    # via beautifulsoup4
sqlparse==0.5.5
    # via django
strawberry-graphql==0.289.3
    # via strawberry-graphql-django
strawberry-graphql-django==0.74.1
    # via computational-finance (pyproject.toml)
tenacity==9.1.4
    # via google-genai
tqdm==4.67.1
    # via computational-finance (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   cross-web
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   strawberry-graphql
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
uuid6==2025.0.1
    # via computational-finance (pyproject.toml)
websockets==16.0
    # via
    #   google-genai
    #   yfinance
xlrd==2.0.2
    # via computational-finance (pyproject.toml)
yfinance==1.1.0
    # via computational-finance (pyproject.toml)
